Add tip about and examples for only reading start of file
An image can be several megabytes in size but the Exif info will be at the
beginning of the file. Therefore we only need to read a small portion of the
image data. Add tip about this in the README file and update the examples with
how to do this.